fixed the saved session
